Вся требуемая логика реализована. Все постман тесты успешно пройдены.

+++ main-service/src/main/java/ru/practicum/mainservice/repository/CategoriesRepository.java
@@ -1,7 +1,7 @@
 package ru.practicum.mainservice.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.practicum.mainservice.model.Category;
+import ru.practicum.mainservice.data.model.Category;
 
-public interface CategoriesRepository extends JpaRepository<Category, Integer> {
+public interface CategoriesRepository extends JpaRepository<Category, Long> {
 }
